+++ DesktopApplicationTemplate.UI/Helpers/AsyncRelayCommand.cs
@@ -26,4 +26,42 @@ namespace DesktopApplicationTemplate.UI.ViewModels
 
         public void RaiseCanExecuteChanged() => CanExecuteChanged?.Invoke(this, EventArgs.Empty);
     }
+
+    /// <summary>
+    /// An <see cref="ICommand"/> implementation that executes asynchronous delegates with a parameter.
+    /// </summary>
+    /// <typeparam name="T">Type of parameter passed to the command.</typeparam>
+    public class AsyncRelayCommand<T> : ICommand
+    {
+        private readonly Func<T?, Task> _execute;
+        private readonly Predicate<T?>? _canExecute;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="AsyncRelayCommand{T}"/> class.
+        /// </summary>
+        public AsyncRelayCommand(Func<T?, Task> execute, Predicate<T?>? canExecute = null)
+        {
+            _execute = execute ?? throw new ArgumentNullException(nameof(execute));
+            _canExecute = canExecute;
+        }
+
+        /// <inheritdoc />
+        public bool CanExecute(object? parameter) => _canExecute?.Invoke((T?)parameter) ?? true;
+
+        /// <inheritdoc />
+        public async void Execute(object? parameter) => await ExecuteAsync((T?)parameter);
+
+        /// <summary>
+        /// Executes the command asynchronously with the provided parameter.
+        /// </summary>
+        public Task ExecuteAsync(T? parameter) => _execute(parameter);
+
+        /// <inheritdoc />
+        public event EventHandler? CanExecuteChanged;
+
+        /// <summary>
+        /// Notifies that the ability to execute has changed.
+        /// </summary>
+        public void RaiseCanExecuteChanged() => CanExecuteChanged?.Invoke(this, EventArgs.Empty);
+    }
 }